🎨 style: [#97] member관련 타입  분리
